+++ apps/frontend-pwa/src/pages/course/[courseId]/index.tsx
@@ -14,11 +14,9 @@ import { twMerge } from 'tailwind-merge'
 import Layout from '@components/Layout'
 import { addApolloState, initializeApollo } from '@lib/apollo'
 import { getParticipantToken } from '@lib/token'
+import { any } from 'ramda'
 
 const { serverRuntimeConfig } = getConfig()
-
-const PLACEHOLDER_IMG =
-  'https://sos-ch-dk-2.exo.io/klicker-uzh-dev/avatars/placeholder.png'
 
 interface ParticipantProps {
   avatar?: string
@@ -48,7 +46,9 @@ function Participant({
         <div className="bg-white border rounded-full">
           <Image
             className="rounded-full"
-            src={`${process.env.NEXT_PUBLIC_AVATAR_BASE_PATH}/${avatar}.svg`}
+            src={`${process.env.NEXT_PUBLIC_AVATAR_BASE_PATH}/${
+              avatar ?? 'placeholder'
+            }.svg`}
             alt=""
             height={30}
             width={30}
@@ -111,7 +111,7 @@ function CourseOverview({ courseId }: any) {
     ],
   })
 
-  const { rank1, rank2, rank3 } = useMemo(() => {
+  const { rank1, rank2, rank3, isSelfContained } = useMemo(() => {
     if (!data?.getCourseOverviewData?.leaderboard) return {}
     return {
       rank1:
@@ -123,6 +123,11 @@ function CourseOverview({ courseId }: any) {
       rank3:
         data.getCourseOverviewData.leaderboard.length >= 3 &&
         data.getCourseOverviewData.leaderboard[2],
+      isSelfContained: any(
+        (item) =>
+          item.participantId === data.getCourseOverviewData?.participant.id,
+        data.getCourseOverviewData.leaderboard
+      ),
     }
   }, [data?.getCourseOverviewData?.leaderboard])
 
@@ -148,7 +153,7 @@ function CourseOverview({ courseId }: any) {
               <ParticipantOther
                 className="bg-white shadow outline-uzh-red-100"
                 pseudonym={rank2.username ?? 'Frei'}
-                avatar={rank2.avatar ?? 'placeholder'}
+                avatar={rank2.avatar}
                 points={rank2.score ?? 0}
               />
             </div>
@@ -160,7 +165,7 @@ function CourseOverview({ courseId }: any) {
               <ParticipantOther
                 className="bg-white shadow outline-uzh-red-100"
                 pseudonym={rank1.username ?? 'Frei'}
-                avatar={rank1.avatar ?? 'placeholder'}
+                avatar={rank1.avatar}
                 points={rank1.score ?? 0}
               />
             </div>
@@ -172,7 +177,7 @@ function CourseOverview({ courseId }: any) {
               <ParticipantOther
                 className="bg-white shadow outline-uzh-red-100"
                 pseudonym={rank3.username ?? 'Frei'}
-                avatar={rank3.avatar ?? 'placeholder'}
+                avatar={rank3.avatar}
                 points={rank3.score ?? 0}
               />
             </div>
@@ -204,10 +209,10 @@ function CourseOverview({ courseId }: any) {
               )
             })}
 
-            {!participation?.isActive && (
+            {(!participation?.isActive || !isSelfContained) && (
               <ParticipantSelf
                 key={participant?.id}
-                isActive={false}
+                isActive={participation.isActive}
                 pseudonym={participant?.username}
                 avatar={participant?.avatar}
                 points={null}